requirements.txt
fastapi
uvicorn[standard]
redis
rq
docx2txt
pypdf
py-pdf-parser
pydantic
python-multipart
pdfplumber
pandas
spacy
sentence-transformers
openai
pinecone-client
streamlit>=1.34
httpx
python-dotenv
docx2txt
spacy[transformers]
rapidfuzz
yake
python-dateutil
nltk